move functions from utilities to erddapy
